# Refactor `lines` command