--- a/crates/nu-command/src/filters/lines.rs
+++ b/crates/nu-command/src/filters/lines.rs
@@ -37,10 +37,6 @@ impl Command for Lines {
 
         let span = input.span().unwrap_or(call.head);
         match input {
-            #[allow(clippy::needless_collect)]
-            // Collect is needed because the string may not live long enough for
-            // the Rc structure to continue using it. If split could take ownership
-            // of the split values, then this wouldn't be needed
             PipelineData::Value(Value::String { val, .. }, ..) => {
                 let mut lines = val.lines().map(String::from).collect::<Vec<_>>();