@fdncred 2024-01-30
This really seems like a great simplification. The only thing I can't see really is if there have been any changes in regard to streaming. We definitely don't want to change how lines streams data if possible.