@IanManske 2024-01-30
> We definitely don't want to change how lines streams data if possible.

Yep, that's for sure. I made sure to prove to myself that the streaming behavior should be the same. Correct me if anything looks wrong:
- for `Value::String`: this already had a `.collect::<Vec<String>>();`, so the whole result is already available and there is no point in streaming.
- for `PipelineData::ListStream`: still streams, just changed how the lines are split for each item.
- for `PipelineData::ExternalStream`: this one might be a little harder to see, but all the logic is still the same. (One small difference is that the incomplete line is removed from the queue instead of the current batch of lines, but this gives the same result.)